Create & Update Product

+++ Backend/Controllers/imageController.js
@@ -14,28 +14,34 @@ cloudinary.config({
 });
 
 const imageController = (req, res, next) => {
-   upload.single("image")(req, res, async (err) => {
-      if (err) {
-         return res.status(400).json({ message: "Image upload failed" });
-      }
+   if (req.body.image && req.body.image.startsWith("http")) {
+      req.imageUrl = req.body.image;
+      next();
+      
+   } else {
+      upload.single("image")(req, res, async (err) => {
+         if (err) {
+            return res.status(400).json({ message: "Image upload failed" });
+         }
 
-      try {
-         const result = await cloudinary.uploader.upload(req.file.path, { folder: "products" });
-         req.imageUrl = result.secure_url;
+         try {
+            const result = await cloudinary.uploader.upload(req.file.path, { folder: "products" });
+            req.imageUrl = result.secure_url;
 
-         fs.unlink(req.file.path, (error) => {
-            if (error) {
-               console.error("Error deleting file:", error);
-            } else {
-               console.log("File deleted successfully");
-            }
-         });
+            fs.unlink(req.file.path, (error) => {
+               if (error) {
+                  console.error("Error deleting file:", error);
+               } else {
+                  console.log("File deleted successfully");
+               }
+            });
 
-         next();
-      } catch (err) {
-         return res.status(500).json({ message: "Failed to upload image to Cloudinary" });
-      }
-   });
+            next();
+         } catch (err) {
+            return res.status(500).json({ message: "Failed to upload image to Cloudinary" });
+         }
+      });
+   }
 };
 
 module.exports = imageController;
